refactor: home timetable #124
