+++ lib/app.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
-import 'dart:convert';
 
+import 'package:dotto/feature/my_page/feature/timetable/controller/timetable_controller.dart';
+import 'package:dotto/feature/my_page/feature/timetable/repository/timetable_repository.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:uni_links/uni_links.dart';
 
@@ -9,13 +10,12 @@ import 'package:dotto/components/color_fun.dart';
 import 'package:dotto/components/setting_user_info.dart';
 import 'package:dotto/repository/db_config.dart';
 import 'package:dotto/repository/download_file_from_firebase.dart';
-import 'package:dotto/repository/narrowed_lessons.dart';
 import 'package:dotto/feature/map/map.dart';
 import 'package:dotto/feature/map/controller/map_controller.dart';
 import 'package:dotto/feature/map/repository/map_repository.dart';
 import 'package:dotto/feature/kamoku_search/kamoku_search.dart';
 import 'package:dotto/screens/app_tutorial.dart';
-import 'package:dotto/screens/home.dart';
+import 'package:dotto/feature/my_page/home.dart';
 import 'package:dotto/screens/kadai_list.dart';
 import 'package:dotto/screens/settings.dart';
 
@@ -169,6 +169,11 @@ class _BasePageState extends ConsumerState<BasePage> {
     return returnParam;
   }
 
+  Future<void> setPersonalLessonIdList() async {
+    ref.read(personalLessonIdListProvider.notifier).state =
+        await TimetableRepository().loadPersonalTimeTableList();
+  }
+
   @override
   void initState() {
     super.initState();
@@ -176,6 +181,7 @@ class _BasePageState extends ConsumerState<BasePage> {
       await initUniLinks();
       await SyllabusDBConfig.setDB();
       await downloadFiles();
+      await setPersonalLessonIdList();
     });
   }
 
@@ -192,17 +198,6 @@ class _BasePageState extends ConsumerState<BasePage> {
     );
   }
 
-  Future<void> loadTimeTableList() async {
-    final jsonString =
-        await UserPreferences.getString(UserPreferenceKeys.kadaiFinishList);
-    if (jsonString != null) {
-      final personalLessonIdListNotifier =
-          ref.watch(personalLessonIdListProvider.notifier);
-      personalLessonIdListNotifier.state =
-          List<int>.from(json.decode(jsonString));
-    }
-  }
-
   Future<void> downloadFiles() async {
     // Firebaseからファイルをダウンロード
     List<String> filePaths = [
@@ -211,7 +206,9 @@ class _BasePageState extends ConsumerState<BasePage> {
       'map/oneweek_schedule.json',
     ];
     for (var path in filePaths) {
+      print("download $path");
       await downloadFileFromFirebase(path);
+      print("finish $path");
     }
   }
