+++ lib/feature/my_page/home.dart
@@ -1,0 +1,207 @@
+import 'package:dotto/components/animation.dart';
+import 'package:dotto/feature/my_page/feature/timetable/controller/timetable_controller.dart';
+import 'package:dotto/feature/my_page/feature/timetable/my_page_timetable.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import 'package:dotto/importer.dart';
+import 'package:dotto/components/color_fun.dart';
+import 'package:dotto/screens/file_viewer.dart';
+import 'package:dotto/feature/my_page/feature/timetable/course_cancellation.dart';
+import 'package:dotto/feature/my_page/feature/timetable/personal_time_table.dart';
+
+class HomeScreen extends ConsumerStatefulWidget {
+  const HomeScreen({super.key});
+
+  @override
+  ConsumerState<HomeScreen> createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends ConsumerState<HomeScreen> {
+  User? currentUser = FirebaseAuth.instance.currentUser;
+  List<int> personalTimeTableList = [];
+
+  void launchUrlInExternal(Uri url) async {
+    if (await canLaunchUrl(url)) {
+      launchUrl(url, mode: LaunchMode.externalApplication);
+    } else {
+      throw 'Could not launch $url';
+    }
+  }
+
+  Widget infoTile(List<Widget> children) {
+    final length = children.length;
+    return Padding(
+      padding: const EdgeInsets.all(5),
+      child: Column(
+        children: [
+          for (int i = 0; i < length; i += 3) ...{
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                for (int j = i; j < i + 3 && j < length; j++) ...{
+                  children[j],
+                }
+              ],
+            ),
+          },
+        ],
+      ),
+    );
+  }
+
+  Widget infoButton(BuildContext context, void Function() onPressed,
+      IconData icon, String title) {
+    final double width = MediaQuery.sizeOf(context).width * 0.26;
+    const double height = 100;
+    return Container(
+      margin: const EdgeInsets.all(5),
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          foregroundColor: Colors.black,
+          surfaceTintColor: Colors.white,
+          fixedSize: Size(width, height),
+        ),
+        onPressed: onPressed,
+        child: Container(
+          padding: const EdgeInsets.symmetric(vertical: 10),
+          width: width,
+          height: height,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              ClipOval(
+                  child: Container(
+                width: 45,
+                height: 45,
+                color: customFunColor,
+                child: Center(
+                    child: Icon(
+                  icon,
+                  color: Colors.white,
+                  size: 25,
+                )),
+              )),
+              const SizedBox(height: 5),
+              Text(
+                title,
+                style: const TextStyle(fontSize: 11),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final double infoBoxWidth = MediaQuery.sizeOf(context).width * 0.4;
+
+    const Map<String, String> fileNamePath = {
+      'バス時刻表': 'home/hakodatebus55.pdf',
+      '学年暦': 'home/academic_calendar.pdf',
+      '前期時間割': 'home/timetable_first.pdf',
+      '後期時間割': 'home/timetable_second.pdf',
+    };
+    List<Widget> infoTiles = [];
+    infoTiles.add(infoButton(context, () {
+      Navigator.of(context).push(
+        PageRouteBuilder(
+          pageBuilder: (context, animation, secondaryAnimation) {
+            return const CourseCancellationScreen();
+          },
+          transitionsBuilder: fromRightAnimation,
+        ),
+      );
+    }, Icons.event_busy, '休講情報'));
+    infoTiles.addAll(fileNamePath.entries
+        .map((item) => infoButton(context, () {
+              Navigator.of(context).push(
+                PageRouteBuilder(
+                  pageBuilder: (context, animation, secondaryAnimation) {
+                    return FileViewerScreen(
+                        filename: item.key,
+                        url: item.value,
+                        storage: StorageService.firebase);
+                  },
+                  transitionsBuilder: fromRightAnimation,
+                ),
+              );
+            }, Icons.picture_as_pdf, item.key))
+        .toList());
+
+    return Scaffold(
+      body: SingleChildScrollView(
+        child: Center(
+          child: Column(
+            children: [
+              const SizedBox(
+                height: 5,
+              ),
+              // 時間割
+              const MyPageTimetable(),
+              Align(
+                alignment: Alignment.centerRight,
+                child: TextButton(
+                  style: TextButton.styleFrom(
+                    padding: const EdgeInsets.symmetric(horizontal: 5),
+                  ),
+                  onPressed: () {
+                    Navigator.of(context)
+                        .push(
+                      PageRouteBuilder(
+                        pageBuilder: (context, animation, secondaryAnimation) {
+                          return const PersonalTimeTableScreen();
+                        },
+                        transitionsBuilder: fromRightAnimation,
+                      ),
+                    )
+                        .then((value) {
+                      ref.invalidate(twoWeekTimeTableDataProvider);
+                    });
+                  },
+                  child: Text(
+                    "時間割を設定する ⇀",
+                    style: TextStyle(
+                      color: Colors.blue.shade700,
+                      decoration: TextDecoration.underline,
+                    ),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 20),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                children: [
+                  ElevatedButton(
+                    onPressed: () async {
+                      const formUrl = 'https://forms.gle/ruo8iBxLMmvScNMFA';
+                      final url = Uri.parse(formUrl);
+                      launchUrlInExternal(url);
+                    },
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: Colors.blue,
+                      fixedSize: Size(infoBoxWidth, 80),
+                      foregroundColor: Colors.white,
+                    ),
+                    child: const Text(
+                      '意見要望\nお聞かせください！',
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        fontSize: 10,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+              const SizedBox(height: 20),
+              infoTile(infoTiles),
+              const SizedBox(height: 20),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
